requirements.txt
beautifulsoup4==4.10.0
Cartopy==0.22.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
GDAL==3.8.4
lxml==4.8.0
matplotlib==3.5.1
numpy==1.21.5
requests==2.25.1
scipy==1.8.0
selenium==4.15.2
shapely==2.0.2
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.10.0
Cartopy==0.22.0
certifi==2024.7.4
chardet==4.0.0
cycler==0.12.1
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
fonttools==4.53.1
GDAL==3.8.4
h11==0.14.0
idna==2.10
kiwisolver==1.4.5
lxml==4.8.0
matplotlib==3.5.1
numpy==1.21.5
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
PyJWT==2.8.0
pyparsing==3.1.2
pyproj==3.6.1
pyshp==2.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.25.1
scipy==1.8.0
selenium==4.15.2
shapely==2.0.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==1.26.19
wsproto==1.2.0
